fix(distribution): move routing js into scripts/version/
